% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{describe}
\alias{describe}
\title{Describe morphology using summary functions}
\usage{
describe(
  .list,
  what = mean,
  of = "neighbours",
  in_function_of = switch(of, neighbours = "distance", distance = "neighbours",
    pathlength = "distance"),
  at = NULL,
  connected = NULL,
  cumulative = FALSE,
  name = NULL
)
}
\arguments{
\item{.list}{List with the \code{distance} matrix and \code{within}, and potentially
a \code{connected} matrix, typically output from \code{look_for()}. Alternatively,
this can be a data frame resulting from a \code{describe()} including a \code{name}
argument, to add additional described summaries.}

\item{what}{Summary function.}

\item{of}{Quantity to summarise, and return as \code{y}. Can be \code{"neighbours"}
(default), \code{"distance"} or \code{"pathlength"}.}

\item{in_function_of}{Quantity to return as \code{x}. The default is to use
\code{"distance"} in case \code{of} is \code{"neighbours"} or \code{"pathlength"}, and
\code{"neighbours"} otherwise.}

\item{at}{Distances at which to describe the results. Defaults to the integer
sequence from \code{0L} to the maximum search distance, if available, and
otherwise the maximum distance found.}

\item{connected}{Logical. Count connected or disconnected neighbours only.
Defaults to \code{NULL}, for which all neighbours are considered.}

\item{cumulative}{Logical. Provide cumulative results (distances <=
threshold) or binned results (distances at center of provided breaks;
default).}

\item{name}{Name to use for the current described dataset in an extra name
column. If provided, the input list is included as attribute to the output,
and \code{describe()} can be repeated multiple times in a pipechain.}
}
\value{
Data frame with \code{distance} and \code{neighbours} columns. If \code{name} is
provided, an additional \code{name} column is included for identifying different
described summaries. In case a \code{name} column is there, the object has an
extra \code{morphology.list} attribute, to enable piping into additional
\code{describe()} calls.
}
\description{
Describe morphology using summary functions
}
