% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{finalise}
\alias{finalise}
\title{Remove all {morphology} object attributes to allow garbage collection}
\usage{
finalise(.x)
}
\arguments{
\item{.x}{Object to remove the attributes from.}
}
\value{
Same object, but without the {morphology} attributes.
}
\description{
Remove all {morphology} object attributes to allow garbage collection
}
