% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_fully_penetrable_spheres}
\alias{gen_fully_penetrable_spheres}
\title{Fully penetrable spheres generator}
\usage{
gen_fully_penetrable_spheres(
  dimensions = rep(200, 3),
  proportion = 0.5,
  radius = 15
)
}
\arguments{
\item{dimensions}{Dimensions of the generated array.}

\item{proportion}{Proportion of the spheres category in the generated array.}

\item{radius}{Radius of the spheres.}
}
\value{
Array with fully penetrable spheres
}
\description{
Fully penetrable spheres generator
}
