% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{look_at}
\alias{look_at}
\title{Define what category to look at}
\usage{
look_at(
  .array,
  what = "voxels",
  of_category,
  in_relation_to = NULL,
  kernel_width = 3,
  kernel_shape = "diamond"
)
}
\arguments{
\item{.array}{Array to analyze.}

\item{what}{Either \code{"voxels"} (default) or \code{"components"} to look at
individual voxels or connected components.}

\item{of_category}{Vector of categories to consider. If negative, these are
omitted.}

\item{in_relation_to}{Vector of categories to consider for cross-category
morphology. If negative, these are omitted.}

\item{kernel_width}{Width of the kernel for determining connectivity. Can be
a vector (row, column, layer). Not used when looking at voxels.}

\item{kernel_shape}{Shape of the kernel for determining connectivity. Can be
\code{"diamond"} (default), \code{"disc"} or \code{"box"}. Not used when looking at
voxels.}
}
\value{
Data frame (tibble) with \code{x}, \code{y}, \code{z} and \code{value} or \code{component}
columns.
}
\description{
Define what category to look at
}
