% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{look_for}
\alias{look_for}
\title{Define what neighbours to look for}
\usage{
look_for(.df, neighbours = 1, within = Inf, from_border = FALSE, error = 0)
}
\arguments{
\item{.df}{Data frame with \code{x}, \code{y}, and/or \code{z}, and \code{value} or \code{component}
columns, typically output from \code{look_at()} or \code{look_in()}.}

\item{neighbours}{Number of nearest neighbours to look for. Can be \code{Inf} if
\code{within} is finite, in which case all neighbours within that distance are
looked for.}

\item{within}{Maximum search distance.}

\item{from_border}{Look for neighbours of border voxels. Defaults to TRUE. If
FALSE, minus sampling is performed.}

\item{error}{Approximate error bound, for approximate nearest neighbour
search.}
}
\value{
List with the \code{distance} matrix and \code{xyz} data frame. When looking at
components, additionally a \code{connected} matrix.
}
\description{
Define what neighbours to look for
}
