% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{look_in}
\alias{look_in}
\title{Define what direction to look in}
\usage{
look_in(.df, direction = "xyz", every = 1)
}
\arguments{
\item{.df}{Data frame with \code{x}, \code{y}, \code{z} and \code{value} or \code{component} columns,
typically output from \code{look_at()}.}

\item{direction}{Direction to consider. Either 1D (\code{"x"}, \code{"y"}, \code{"z"}), 2D
(\code{"xy"}, \code{"xz"}, \code{"yz"}) or 3D (\code{"xyz"}).}

\item{every}{Value indicating how many voxels to consider, e.g. one of two
(\code{1/2}), one every three (\code{1/3}), etc. Defaults to every voxel (\code{1/1}).
This can be a vector (row, column, layer) to thin the different dimensions
differently. Alternatively it can be a list of two vectors, where the first
one is used for the category of interest (\code{of_category} in \code{look_at()}) and
the second one is for the neighbour category (\code{in_relation_to} in
\code{look_at()}, or again \code{of_category} for a self-query).}
}
\value{
Data frame (tibble) with \code{x}, \code{y}, and/or \code{z}, and \code{value} or
\code{component} columns.
}
\description{
Define what direction to look in
}
