% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{q05}
\alias{q05}
\title{Calculate the 0.05 quantile}
\usage{
q05(...)
}
\arguments{
\item{...}{Arguments passed to the \code{quantile()} function.}
}
\value{
Sample quantile.
}
\description{
Calculate the 0.05 quantile
}
