% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{q25}
\alias{q25}
\title{Calculate the 0.25 quantile}
\usage{
q25(...)
}
\arguments{
\item{...}{Arguments passed to the \code{quantile()} function.}
}
\value{
Sample quantile.
}
\description{
Calculate the 0.25 quantile
}
