% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{q50}
\alias{q50}
\title{Calculate the 0.50 quantile}
\usage{
q50(...)
}
\arguments{
\item{...}{Arguments passed to the \code{quantile()} function.}
}
\value{
Sample quantile.
}
\description{
Calculate the 0.50 quantile
}
