% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{q75}
\alias{q75}
\title{Calculate the 0.75 quantile}
\usage{
q75(...)
}
\arguments{
\item{...}{Arguments passed to the \code{quantile()} function.}
}
\value{
Sample quantile.
}
\description{
Calculate the 0.75 quantile
}
