% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{q95}
\alias{q95}
\title{Calculate the 0.95 quantile}
\usage{
q95(...)
}
\arguments{
\item{...}{Arguments passed to the \code{quantile()} function.}
}
\value{
Sample quantile.
}
\description{
Calculate the 0.95 quantile
}
