% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{scale_by}
\alias{scale_by}
\title{Scale the morphological description}
\usage{
scale_by(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Things to scale the description by. Can be \code{"neighbourhood"},
which automatically chooses between alternative options \code{"volume"},
\code{"area"} or \code{"length"}, or \code{"proportion"}, \code{"inverse proportion"}, another
description data frame with the same distances, ...}
}
\value{
Data frame with scaled \code{neighbours} column.
}
\description{
Scale the morphological description
}
