% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{visualise}
\alias{visualise}
\title{Visualise the morphological description}
\usage{
visualise(df, ...)
}
\arguments{
\item{df}{Data frame resulting from a \code{describe()} call, or just a numeric
vector.}

\item{...}{Arguments passed to \code{ggplot2::labs()} for labelling.}
}
\value{
A ggplot2 plot.
}
\description{
Visualise the morphological description
}
